view cart and cart detail by account. add to cart, add reduce remove product in cart

+++ src/main/java/edu/poly/shop/domain/OrderDetailDto.java
@@ -13,7 +13,7 @@ public class OrderDetailDto {
 
     private int orderDetailid;
     private int orderid;
-    private int productid;
+    private Long productid;
     private int quantity;
     private double unitPrice;
